+++ UI/Models/BuildConfiguration.cs
@@ -53,14 +53,14 @@
             stringBuilder.AppendLine(@"#---------------------------------- Content ---------------------------------#");
             stringBuilder.AppendLine();
 
-            var gameSettingsPath = $@"{projectDirectoryPath}\{GameSettings.ContentFileName}{FileHelper.GameSettingsExtension}";
+            var gameSettingsPath = $@"{GameSettings.ContentFileName}{FileHelper.GameSettingsExtension}";
             stringBuilder.AppendLine($"#begin {gameSettingsPath}");
             stringBuilder.AppendLine($@"/importer:{nameof(GameSettingsImporter)}");
             stringBuilder.AppendLine($@"/processor:{nameof(GameSettingsProcessor)}");
             stringBuilder.AppendLine($@"/build:{gameSettingsPath}");
             stringBuilder.AppendLine();
 
-            var assetManagerPath = $@"{projectDirectoryPath}\{AssetManager.ContentFileName}{FileHelper.AssetManagerExtension}";
+            var assetManagerPath = $@"{AssetManager.ContentFileName}{FileHelper.AssetManagerExtension}";
             stringBuilder.AppendLine($"#begin {assetManagerPath}");
             stringBuilder.AppendLine($@"/importer:{nameof(AssetManagerImporter)}");
             stringBuilder.AppendLine($@"/processor:{nameof(AssetManagerProcessor)}");
@@ -68,11 +68,11 @@
             stringBuilder.AppendLine();
 
             foreach (var asset in assets) {
-                asset.BuildProcessorCommands(stringBuilder, projectDirectoryPath);
+                asset.BuildProcessorCommands(stringBuilder);
                 stringBuilder.AppendLine();
             }
 
-            var contentFile = Path.Combine(projectDirectoryPath, isTemp ? $"{FileHelper.TempName}{FileHelper.ContentExtension}" : $"{this.Platform.ToString()}{FileHelper.ContentExtension}");
+            var contentFile = Path.Combine(projectDirectoryPath, "Content", isTemp ? $"{FileHelper.TempName}{FileHelper.ContentExtension}" : $"{this.Platform.ToString()}{FileHelper.ContentExtension}");
             File.WriteAllText(contentFile, stringBuilder.ToString());
         }