+++ UI/Models/FontAsset.cs
@@ -121,8 +121,8 @@
             }
         }
 
-        public override void BuildProcessorCommands(StringBuilder contentStringBuilder, string projectDirectoryPath) {
-            var path = Path.Combine(projectDirectoryPath, this.GetContentPath());
+        public override void BuildProcessorCommands(StringBuilder contentStringBuilder) {
+            var path = this.GetContentPath();
             contentStringBuilder.AppendLine($"#begin {path}");
             contentStringBuilder.AppendLine(@"/importer:FontDescriptionImporter");
             contentStringBuilder.AppendLine(@"/processor:FontDescriptionProcessor");