+++ UI/Models/ImageAsset.cs
@@ -104,8 +104,8 @@
             return wrapper;
         }
 
-        public override void BuildProcessorCommands(StringBuilder contentStringBuilder, string projectDirectoryPath) {
-            var path = Path.Combine(projectDirectoryPath, this.GetContentPath());
+        public override void BuildProcessorCommands(StringBuilder contentStringBuilder) {
+            var path = this.GetContentPath();
             contentStringBuilder.AppendLine($"#begin {path}");
             contentStringBuilder.AppendLine(@"/importer:TextureImporter");
             contentStringBuilder.AppendLine(@"/processor:TextureProcessor");