+++ UI/Models/SceneAsset.cs
@@ -86,8 +86,8 @@
             return result;
         }
 
-        public override void BuildProcessorCommands(StringBuilder contentStringBuilder, string projectDirectoryPath) {
-            var path = Path.Combine(projectDirectoryPath, this.GetContentPath());
+        public override void BuildProcessorCommands(StringBuilder contentStringBuilder) {
+            var path = this.GetContentPath();
             contentStringBuilder.AppendLine($"#begin {path}");
             contentStringBuilder.AppendLine(@"/importer:SceneImporter");
             contentStringBuilder.AppendLine(@"/processor:SceneProcessor");