+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * All asset types are properly synced.
     * This means that when editing auto tile sets and sprite animations, auto tile maps and sprite animators will be updated appropriately.
 * Auto tile sets now sync properly with their sprites in the editor.
+* Content should build in both the editor and the game project.
 
 ## Compatibility Breaks